+++ src/Termonad/Config.hs
@@ -61,5 +61,5 @@ module Termonad.Config
 
 import GI.Vte (CursorBlinkMode(..))
 
-import Termonad.PreferencesFile (tmConfigFromPreferencesFile)
+import Termonad.Preferences (tmConfigFromPreferencesFile)
 import Termonad.Types